requirements.txt
opencv-python
param
Pillow
panel